Converting additionalProps to primitives
