+++ components/twitter/actions/get-tweet/get-tweet.ts
@@ -1,9 +1,7 @@
 import app from "../../app/twitter.app";
 import { defineAction } from "@pipedream/types";
 import { GetTweetParams } from "../../common/types/requestParams";
-import {
-  includeAllFields, tweetAdditionalProps as additionalProps,
-} from "../../common/propGroups";
+import { tweetAdditionalProps as additionalProps } from "../../common/additionalProps";
 import { getTweetFields } from "../../common/methods";
 import {
   ResponseObject, Tweet,
@@ -16,7 +14,7 @@ export default defineAction({
   key: "twitter-get-tweet",
   name: "Get Tweet",
   description: `Return a single tweet specified by ID. [See docs here](${DOCS_LINK})`,
-  version: "1.0.0",
+  version: "1.0.1",
   type: "action",
   props: {
     app,
@@ -26,22 +24,28 @@ export default defineAction({
         "tweetId",
       ],
     },
-    includeAllFields,
+    includeAllFields: {
+      propDefinition: [
+        app,
+        "includeAllFields",
+      ],
+    },
   },
   additionalProps,
   methods: {
     getTweetFields,
   },
   async run({ $ }): Promise<ResponseObject<Tweet>> {
+    const { tweetId } = this;
     const params: GetTweetParams = {
       $,
       params: this.getTweetFields(),
-      tweetId: this.tweetId,
+      tweetId,
     };
 
     const response = await this.app.getTweet(params);
 
-    $.export("$summary", "Successfully retrieved tweet");
+    $.export("$summary", `Successfully retrieved tweet ${tweetId}`);
 
     return response;
   },
