+++ components/twitter/sources/new-list-followed/new-list-followed.ts
@@ -2,9 +2,7 @@ import app from "../../app/twitter.app";
 import { defineSource } from "@pipedream/types";
 import common from "../common/base";
 import { getListSummary as getItemSummary } from "../common/getItemSummary";
-import {
-  includeAllFields, listAdditionalProps as additionalProps,
-} from "../../common/propGroups";
+import { listAdditionalProps as additionalProps } from "../../common/additionalProps";
 import { List } from "../../common/types/responseSchemas";
 import { getListFields } from "../../common/methods";
 import { GetUserFollowedListsParams } from "../../common/types/requestParams";
@@ -28,7 +26,12 @@ export default defineSource({
         "userNameOrId",
       ],
     },
-    includeAllFields,
+    includeAllFields: {
+      propDefinition: [
+        app,
+        "includeAllFields",
+      ],
+    },
   },
   additionalProps,
   methods: {
